+++ packages/atlas-ui/index.ts
@@ -11,6 +11,7 @@ export { default as AtlasBadge } from '../../src/components/ui/AtlasBadge.vue'
 export { default as AtlasBanner } from '../../src/components/ui/AtlasBanner.vue'
 export { default as AtlasButton } from '../../src/components/ui/AtlasButton.vue'
 export { default as AtlasCard } from '../../src/components/ui/AtlasCard.vue'
+export { default as AtlasChartExport } from '../../src/components/ui/charts/AtlasChartExport.vue'
 export { default as AtlasCheckbox } from '../../src/components/ui/AtlasCheckbox.vue'
 export { default as AtlasChip } from '../../src/components/ui/AtlasChip.vue'
 export { default as AtlasCol } from '../../src/components/ui/AtlasCol.vue'
