+++ src/components/reports/charts/BarChart.vue
@@ -39,7 +39,7 @@ import { computed, ref, onMounted, onUnmounted, watch } from 'vue'
 import type { BarChartData } from '@/models/report.types'
 import type { EChartsType } from 'echarts/core'
 import { defaultBarChartOptions, createResizeHandler } from '@/utils/chart-config'
-import ChartExport from './ChartExport.vue'
+import ChartExport from '@/components/ui/charts/AtlasChartExport.vue'
 
 /**
  * Props
