+++ src/components/reports/charts/BoxPlotChart.vue
@@ -39,7 +39,7 @@ import { computed, ref, onMounted, onUnmounted, watch } from 'vue'
 import type { BoxPlotData } from '@/models/report.types'
 import type { EChartsType } from 'echarts/core'
 import { boxPlotChartOptions, createResizeHandler } from '@/utils/chart-config'
-import ChartExport from './ChartExport.vue'
+import ChartExport from '@/components/ui/charts/AtlasChartExport.vue'
 
 /**
  * Props
